+++ lib/src/services/yust_database_service_dart.dart
@@ -186,7 +186,7 @@ class YustDatabaseService {
         'getFirstFromDB',
         _getDocumentPath(docSetup),
         () => _api.projects.databases.documents.runQuery(
-            _getQuery(docSetup, filters: filters, orderBy: orderBy, limit: 1),
+            getQuery(docSetup, filters: filters, orderBy: orderBy, limit: 1),
             _getParentPath(docSetup)));
 
     if (response.isEmpty || response.first.document == null) {
@@ -276,7 +276,7 @@ class YustDatabaseService {
         'getListFromDB',
         _getDocumentPath(docSetup),
         () => _api.projects.databases.documents.runQuery(
-            _getQuery(docSetup,
+            getQuery(docSetup,
                 filters: filters, orderBy: orderBy, limit: limit),
             _getParentPath(docSetup)));
     dbLogCallback?.call(
@@ -341,7 +341,7 @@ class YustDatabaseService {
       var isDone = false;
       String? lastDocument;
       while (!isDone) {
-        final request = _getQuery(docSetup,
+        final request = getQuery(docSetup,
             filters: filters,
             // orderBy __name__ is required for pagination
             orderBy: [...?orderBy, YustOrderBy(field: '__name__')],
@@ -520,7 +520,7 @@ class YustDatabaseService {
     // Because the Firestore REST-Api (used in the background) can't handle attributes starting with numbers,
     // e.g. 'foo.0bar', we need to escape the path-parts by using '´': '`foo`.`0bar`'
     final quotedUpdateMask = updateMask
-        ?.map((path) => YustHelpers().toQuotedFieldPath(path))
+        ?.map((path) => YustHelpers().toQuotedFieldPath(path)!)
         .toList();
     final docPath = _getDocumentPath(docSetup, doc.id);
     await _retryOnException('saveDoc', _getDocumentPath(docSetup), () async {
@@ -591,7 +591,7 @@ class YustDatabaseService {
     int? limit,
   }) async {
     final response = await _api.projects.databases.documents.runQuery(
-      _getQuery(docSetup, filters: filters, orderBy: orderBy, limit: limit),
+      getQuery(docSetup, filters: filters, orderBy: orderBy, limit: limit),
       _getParentPath(docSetup),
     );
 
@@ -782,7 +782,7 @@ class YustDatabaseService {
       write.updateMask = DocumentMask(
           fieldPaths: doc.updateMask
               .map(
-                (e) => YustHelpers().toQuotedFieldPath(e),
+                (e) => YustHelpers().toQuotedFieldPath(e)!,
               )
               .toList());
     }
@@ -805,17 +805,6 @@ class YustDatabaseService {
         () => _api.projects.databases.documents
             .commit(commitRequest, _getDatabasePath()),
         shouldRetryOnTransactionErrors: false);
-  }
-
-  /// Returns a query for specified filter and order.
-  dynamic getQueryWithLogging<T extends YustDoc>(
-    YustDocSetup<T> docSetup, {
-    List<YustFilter>? filters,
-    List<YustOrderBy>? orderBy,
-    int? limit,
-  }) {
-    return _getQuery<T>(docSetup,
-        filters: filters, orderBy: orderBy, limit: limit);
   }
 
   /// Transforms a json to a [YustDoc]
@@ -853,7 +842,7 @@ class YustDatabaseService {
     return '${_getParentPath(docSetup)}/${_getCollection(docSetup)}/$id';
   }
 
-  RunQueryRequest _getQuery<T extends YustDoc>(
+  RunQueryRequest getQuery<T extends YustDoc>(
     YustDocSetup<T> docSetup, {
     List<YustFilter>? filters,
     List<YustOrderBy>? orderBy,
@@ -885,6 +874,7 @@ class YustDatabaseService {
     final countAggregation = Aggregation(
         alias: AggregationType.count.name,
         count: Count(upTo: upTo?.toString()));
+    final quotedFieldPath = YustHelpers().toQuotedFieldPath(fieldPath);
     return RunAggregationQueryRequest(
       structuredAggregationQuery: StructuredAggregationQuery(
         aggregations: [
@@ -894,12 +884,12 @@ class YustDatabaseService {
             Aggregation(
                 alias: type.name,
                 count: Count(upTo: upTo?.toString()),
-                sum: Sum(field: FieldReference(fieldPath: fieldPath))),
+                sum: Sum(field: FieldReference(fieldPath: quotedFieldPath))),
           if (type == AggregationType.avg)
             Aggregation(
                 alias: type.name,
                 count: Count(upTo: upTo?.toString()),
-                avg: Avg(field: FieldReference(fieldPath: fieldPath))),
+                avg: Avg(field: FieldReference(fieldPath: quotedFieldPath))),
         ],
         structuredQuery: StructuredQuery(
           from: [CollectionSelector(collectionId: _getCollection(docSetup))],
@@ -996,7 +986,8 @@ class YustDatabaseService {
   List<Order> _executeOrderByList(List<YustOrderBy>? orderBy) {
     return orderBy
             ?.map((e) => Order(
-                field: FieldReference(fieldPath: e.field),
+                field: FieldReference(
+                    fieldPath: YustHelpers().toQuotedFieldPath(e.field)),
                 direction: e.descending ? 'DESCENDING' : 'ASCENDING'))
             .toList() ??
         <Order>[];